Add Ray backend
